Allow to use tokens like {app.name}/{app.version}/... in the mail subject and the mail body.
This allows to show the bug information in a mail as a table instead of just deserialized xml.

An example for a custom body:
Hi, an exception was logged: {exception.type}

      Best regards,
            {app.name} {app.version}

This commit does not allow the user to add custom info. Like a bug ID. Not sure how to do that.